refactor: CategoryBox
Related to : #52
